requirements.txt
-i https://pypi.org/simple
aiodocker==0.24.0; python_full_version >= '3.9.0'
aiohappyeyeballs==2.4.4; python_version >= '3.8'
aiohttp==3.11.10; python_version >= '3.9'
aiosignal==1.3.2; python_version >= '3.9'
annotated-types==0.7.0; python_version >= '3.8'
anthropic==0.42.0; python_version >= '3.8'
anyio==4.7.0; python_version >= '3.9'
async-timeout==5.0.1; python_version >= '3.8'
asyncio==3.4.3
attrs==24.3.0; python_version >= '3.8'
beautifulsoup4==4.12.3; python_full_version >= '3.6.0'
brotli==1.1.0
bs4==0.0.2
certifi==2024.12.14; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
coverage==7.6.9; python_version >= '3.9'
defusedxml==0.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
distro==1.9.0; python_version >= '3.6'
dnspython==2.7.0; python_version >= '3.9'
duckduckgo-search[lxml]==6.3.7; python_version >= '3.8'
environs==11.2.1; python_version >= '3.8'
exceptiongroup==1.2.2; python_version < '3.11'
frozenlist==1.5.0; python_version >= '3.8'
h11==0.14.0; python_version >= '3.7'
hibpwned==1.3.7
httpcore==1.0.7; python_version >= '3.8'
httpx==0.27.2; python_version >= '3.8'
idna==3.10; python_version >= '3.6'
jira==3.8.0; python_version >= '3.8'
jiter==0.8.2; python_version >= '3.8'
jsonpickle==4.0.1; python_version >= '3.8'
libvalkey==4.0.1; python_version >= '3.8'
lxml==5.3.0
marshmallow==3.23.1; python_version >= '3.9'
mattermostautodriver==1.3.0; python_version >= '3.6'
mmpy-bot==2.2.0; python_version >= '3.8'
multidict==6.1.0; python_version >= '3.8'
ntplib==0.4.0
oauthlib==3.2.2; python_version >= '3.6'
openai==1.58.1; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pedantic==1.16.4; python_full_version >= '3.9.0'
pillow==11.0.0; python_version >= '3.9'
primp==0.9.1; python_version >= '3.8'
propcache==0.2.1; python_version >= '3.9'
pydantic==2.10.3; python_version >= '3.8'
pydantic-core==2.27.1; python_version >= '3.8'
python-dotenv==1.0.1; python_version >= '3.8'
python-magic==0.4.27; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
python-redis-rate-limit==0.0.10
pyttsx3==2.98
redis==5.2.1; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-oauthlib==2.0.0; python_version >= '3.4'
requests-toolbelt==1.0.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
schedule==1.2.2; python_version >= '3.7'
sniffio==1.3.1; python_version >= '3.7'
soupsieve==2.6; python_version >= '3.8'
tiktoken==0.8.0; python_version >= '3.9'
tqdm==4.67.1; python_version >= '3.7'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
uuid==1.30
validators==0.34.0; python_version >= '3.8'
valkey[libvalkey]==6.0.2; python_version >= '3.8'
wheel==0.45.1; python_version >= '3.8'
yarl==1.18.3; python_version >= '3.9'
